Port 10 GGMLOps functions: sin/cos, GLU family, set_rows, add_id, repeat_4d, argsort_top_k
- Fix ggmlSin: was incorrectly using ABS op, now properly uses SIN op with inplace variant
- Fix ggmlCos: was error stub, now properly uses COS op with inplace variant
- Rewrite ggmlGlu/GLU family: use proper GGMLOp.GLU (was MAP_UNARY), add ggmlGluImpl
  matching C ggml_glu_impl (ggml.c:2862) with contiguity asserts
- Add all GLU variants: reglu/geglu/swiglu split, swapped, erf, quick, swiglu_oai
- Port ggmlGetTensor: walks GGMLObject linked list matching C (ggml.c:1965)
- Port ggmlAddId: assert shapes + I32 ids, GGML_OP_ADD_ID (ggml.c:2077)
- Port ggmlRepeat4d: assert divisibility, GGML_OP_REPEAT (ggml.c:2540)
- Port ggmlSetRows: preserve unusual src ordering b,c,a (ggml.c:3884)
- Port ggmlArgsortTopK: compose argsort + view4d (ggml.c:5258)
- Remove duplicate old GLU functions (geglu_erf/quick with wrong param encoding)
- Remove duplicate sin/cos inplace functions
- Fix LlamaGraph.kt: nExpertUsed Long->Int for argsort_top_k parameter

+++ src/nativeMain/kotlin/ai/solace/llamakotlin/model/LlamaGraph.kt
@@ -1834,7 +1834,7 @@ open class LlmGraphContext(val params: LlmGraphParams) {
         }
 
         // -- top-k expert selection -----------------------------------------
-        val selectedExperts = ggmlArgsortTopK(c, selectionProbs, nExpertUsed)
+        val selectedExperts = ggmlArgsortTopK(c, selectionProbs, nExpertUsed.toInt())
         cb(selectedExperts, "ffn_moe_topk", il)
 
         probs = ggmlReshape3d(c, probs, 1, nExpert, nTokensCur)
